requirements.txt
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
websocket-client==1.8.0
cryptography==42.0.8
